Bugfix: Magnetometer would malfunction in high-performance modes
Using the high-performance or ultra-high-performance modes would cause the
magnetometer reading to be gibberish. This is because the driver sets the
wrong bits in the registers. In particular, using high-power modes will
cause bit BLE in CTRL_REG4 to be set, changing the endianness of the
reading, which explains the garbage values.

Someone did not pay close attention to the datasheet.